+++ crates/eips/src/eip4844/env_settings.rs
@@ -63,4 +63,13 @@ impl EnvKzgSettings {
             Self::Custom(settings) => settings,
         }
     }
+
+    /// Load custom KZG settings from a trusted setup file.
+    #[cfg(feature = "std")]
+    pub fn load_from_trusted_setup_file(
+        trusted_setup_file: &std::path::Path,
+    ) -> Result<Self, c_kzg::Error> {
+        let settings = KzgSettings::load_trusted_setup_file(trusted_setup_file)?;
+        Ok(Self::Custom(Arc::new(settings)))
+    }
 }