+++ crates/json-rpc/src/result.rs
@@ -62,7 +62,7 @@ where
 {
     let json = result?;
     let json = json.borrow().get();
-    trace!(ty=%std::any::type_name::<T>(), json, "deserializing response");
+    trace!(ty=%std::any::type_name::<T>(), %json, "deserializing response");
     serde_json::from_str(json)
         .inspect(|response| trace!(?response, "deserialized response"))
         .inspect_err(|err| trace!(?err, "failed to deserialize response"))