+++ crates/rpc-types-engine/src/payload.rs
@@ -203,7 +203,7 @@ impl ExecutionPayloadV1 {
 /// See also: <https://github.com/ethereum/execution-apis/blob/6709c2a795b707202e93c4f2867fa0bf2640a84f/src/engine/shanghai.md#executionpayloadv2>
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
-#[cfg_attr(feature = "serde", serde(rename_all = "camelCase", deny_unknown_fields))]
+#[cfg_attr(feature = "serde", serde(rename_all = "camelCase"))]
 pub struct ExecutionPayloadV2 {
     /// Inner V1 payload
     #[cfg_attr(feature = "serde", serde(flatten))]
@@ -446,6 +446,7 @@ impl ssz::Encode for ExecutionPayloadV3 {
 /// This includes all bundled blob related data of an executed payload.
 #[derive(Clone, Debug, PartialEq, Eq)]
 #[cfg_attr(feature = "serde", derive(serde::Serialize, serde::Deserialize))]
+#[cfg_attr(feature = "ssz", derive(ssz_derive::Encode, ssz_derive::Decode))]
 pub struct BlobsBundleV1 {
     /// All commitments in the bundle.
     pub commitments: Vec<alloy_consensus::Bytes48>,
@@ -453,66 +454,6 @@ pub struct BlobsBundleV1 {
     pub proofs: Vec<alloy_consensus::Bytes48>,
     /// All blobs in the bundle.
     pub blobs: Vec<alloy_consensus::Blob>,
-}
-
-#[cfg(feature = "ssz")]
-#[derive(ssz_derive::Encode, ssz_derive::Decode)]
-struct BlobsBundleV1Ssz {
-    commitments: Vec<alloy_primitives::FixedBytes<48>>,
-    proofs: Vec<alloy_primitives::FixedBytes<48>>,
-    blobs: Vec<alloy_primitives::FixedBytes<{ alloy_eips::eip4844::BYTES_PER_BLOB }>>,
-}
-
-#[cfg(feature = "ssz")]
-impl BlobsBundleV1Ssz {
-    const _ASSERT: [(); std::mem::size_of::<BlobsBundleV1>()] = [(); std::mem::size_of::<Self>()];
-
-    const fn wrap_ref(other: &BlobsBundleV1) -> &Self {
-        // SAFETY: Same repr and size
-        unsafe { &*(other as *const BlobsBundleV1 as *const Self) }
-    }
-
-    fn unwrap(self) -> BlobsBundleV1 {
-        BlobsBundleV1 { commitments: self.commitments, proofs: self.proofs, blobs: self.blobs }
-    }
-}
-
-#[cfg(feature = "ssz")]
-impl ssz::Encode for BlobsBundleV1 {
-    fn is_ssz_fixed_len() -> bool {
-        <BlobsBundleV1Ssz as ssz::Encode>::is_ssz_fixed_len()
-    }
-
-    fn ssz_append(&self, buf: &mut Vec<u8>) {
-        BlobsBundleV1Ssz::wrap_ref(self).ssz_append(buf)
-    }
-
-    fn ssz_fixed_len() -> usize {
-        <BlobsBundleV1Ssz as ssz::Encode>::ssz_fixed_len()
-    }
-
-    fn ssz_bytes_len(&self) -> usize {
-        BlobsBundleV1Ssz::wrap_ref(self).ssz_bytes_len()
-    }
-
-    fn as_ssz_bytes(&self) -> Vec<u8> {
-        BlobsBundleV1Ssz::wrap_ref(self).as_ssz_bytes()
-    }
-}
-
-#[cfg(feature = "ssz")]
-impl ssz::Decode for BlobsBundleV1 {
-    fn is_ssz_fixed_len() -> bool {
-        <BlobsBundleV1Ssz as ssz::Decode>::is_ssz_fixed_len()
-    }
-
-    fn ssz_fixed_len() -> usize {
-        <BlobsBundleV1Ssz as ssz::Decode>::ssz_fixed_len()
-    }
-
-    fn from_ssz_bytes(bytes: &[u8]) -> Result<Self, ssz::DecodeError> {
-        BlobsBundleV1Ssz::from_ssz_bytes(bytes).map(BlobsBundleV1Ssz::unwrap)
-    }
 }
 
 impl BlobsBundleV1 {
@@ -1131,90 +1072,6 @@ mod tests {
         "#;
         let payload: ExecutionPayloadInputV2 = serde_json::from_str(response).unwrap();
         assert_eq!(payload.withdrawals, None);
-    }
-
-    #[test]
-    #[cfg(feature = "serde")]
-    fn serde_deserialize_v3_with_unknown_fields() {
-        let input = r#"
-{
-    "parentHash": "0xaaa4c5b574f37e1537c78931d1bca24a4d17d4f29f1ee97e1cd48b704909de1f",
-    "feeRecipient": "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba",
-    "stateRoot": "0x308ee9c5c6fab5e3d08763a3b5fe0be8ada891fa5010a49a3390e018dd436810",
-    "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
-    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "blockNumber": "0xf",
-    "gasLimit": "0x16345785d8a0000",
-    "gasUsed": "0x0",
-    "timestamp": "0x3a97",
-    "extraData": "0x",
-    "baseFeePerGas": "0x7",
-    "blockHash": "0x38bb6ba645c7e6bd970f9c7d492fafe1e04d85349054cb48d16c9d2c3e3cd0bf",
-    "transactions": [],
-    "withdrawals": [],
-    "excessBlobGas": "0x0",
-    "blobGasUsed": "0x0"
-}
-        "#;
-
-        // ensure that deserializing this succeeds
-        let _payload_res: ExecutionPayloadV3 = serde_json::from_str(input).unwrap();
-
-        // construct a payload with a random field in the middle
-        let input = r#"
-{
-    "parentHash": "0xaaa4c5b574f37e1537c78931d1bca24a4d17d4f29f1ee97e1cd48b704909de1f",
-    "feeRecipient": "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba",
-    "stateRoot": "0x308ee9c5c6fab5e3d08763a3b5fe0be8ada891fa5010a49a3390e018dd436810",
-    "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
-    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "blockNumber": "0xf",
-    "gasLimit": "0x16345785d8a0000",
-    "gasUsed": "0x0",
-    "timestamp": "0x3a97",
-    "extraData": "0x",
-    "baseFeePerGas": "0x7",
-    "blockHash": "0x38bb6ba645c7e6bd970f9c7d492fafe1e04d85349054cb48d16c9d2c3e3cd0bf",
-    "transactions": [],
-    "withdrawals": [],
-    "randomStuff": [],
-    "excessBlobGas": "0x0",
-    "blobGasUsed": "0x0"
-}
-        "#;
-
-        // ensure that deserializing this fails
-        let _payload_res = serde_json::from_str::<ExecutionPayloadV3>(input).unwrap_err();
-
-        // construct a payload with a random field at the end
-        let input = r#"
-{
-    "parentHash": "0xaaa4c5b574f37e1537c78931d1bca24a4d17d4f29f1ee97e1cd48b704909de1f",
-    "feeRecipient": "0x2adc25665018aa1fe0e6bc666dac8fc2697ff9ba",
-    "stateRoot": "0x308ee9c5c6fab5e3d08763a3b5fe0be8ada891fa5010a49a3390e018dd436810",
-    "receiptsRoot": "0x56e81f171bcc55a6ff8345e692c0f86e5b48e01b996cadc001622fb5e363b421",
-    "logsBloom": "0x00000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000",
-    "prevRandao": "0x0000000000000000000000000000000000000000000000000000000000000000",
-    "blockNumber": "0xf",
-    "gasLimit": "0x16345785d8a0000",
-    "gasUsed": "0x0",
-    "timestamp": "0x3a97",
-    "extraData": "0x",
-    "baseFeePerGas": "0x7",
-    "blockHash": "0x38bb6ba645c7e6bd970f9c7d492fafe1e04d85349054cb48d16c9d2c3e3cd0bf",
-    "transactions": [],
-    "withdrawals": [],
-    "randomStuff": [],
-    "excessBlobGas": "0x0",
-    "blobGasUsed": "0x0"
-    "moreRandomStuff": "0x0",
-}
-        "#;
-
-        // ensure that deserializing this fails
-        let _payload_res = serde_json::from_str::<ExecutionPayloadV3>(input).unwrap_err();
     }
 
     #[test]