+++ crates/rpc-types-eth/src/log.rs
@@ -1,5 +1,3 @@
-use core::borrow::Borrow;
-
 use alloy_primitives::{Address, BlockHash, LogData, TxHash, B256};
 
 /// Ethereum Log emitted by a transaction
@@ -153,20 +151,21 @@ impl<T> AsMut<T> for Log<T> {
     }
 }
 
-impl<T> Borrow<alloy_primitives::Log<T>> for Log<T> {
-    fn borrow(&self) -> &alloy_primitives::Log<T> {
-        &self.inner
-    }
-}
-
 #[cfg(test)]
 mod tests {
-    use alloy_primitives::{Address, Bytes};
-
     use super::*;
+    use alloy_consensus::{Receipt, ReceiptWithBloom, TxReceipt};
+    use alloy_primitives::{Address, Bytes};
     use arbitrary::Arbitrary;
     use rand::Rng;
     use similar_asserts::assert_eq;
+
+    const fn assert_tx_receipt<T: TxReceipt>() {}
+
+    #[test]
+    const fn assert_receipt() {
+        assert_tx_receipt::<ReceiptWithBloom<Receipt<Log>>>();
+    }
 
     #[test]
     fn log_arbitrary() {